requirements.txt
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.26.1
python-dotenv==0.19.2
selenium==4.7.0

tabulate~=0.9.0
google~=3.0.0